refactored unit test and integration tests
