+++ test/test_app.py
@@ -5,7 +5,7 @@ from unittest import mock
 import httpx
 import pytest
 
-from app.app import _pull_workouts_page, pull_all_workouts, truncate_and_store_workouts
+from app.app import pull_all_workouts, truncate_and_store_workouts
 
 
 @pytest.mark.asyncio
@@ -34,7 +34,7 @@ async def test_pull_all_workouts():
     m_pull_workouts_page.side_effect = test_pull_workout_effect
     with (
         mock.patch("app.app.truncate_and_store_workouts", m_truncate_and_store),
-        mock.patch("app.app._pull_workouts_page", m_pull_workouts_page),
+        mock.patch("app.app._hevy_api_repo._pull_workouts_page", m_pull_workouts_page),
     ):
         await pull_all_workouts(page_size=m_page_size)
 
@@ -50,7 +50,7 @@ async def test_pull_all_workouts_http_error():
         raise httpx.HTTPError("HTTP error")
 
     m_pull_workouts_page = mock.AsyncMock(side_effect=test_pull_workout_effect)
-    with mock.patch("app.app._pull_workouts_page", m_pull_workouts_page):
+    with mock.patch("app.app._hevy_api_repo._pull_workouts_page", m_pull_workouts_page):
         with pytest.raises(httpx.HTTPError, match="HTTP error"):
             await pull_all_workouts()
 
@@ -59,7 +59,7 @@ async def test_pull_all_workouts_http_error():
 async def test_truncate_and_store_workouts():
     """Test truncating and storing workouts."""
     m_db = "hevy_db"
-    with mock.patch("app.app.Mongo_Conn") as m_conn:
+    with mock.patch("app.app.MongoConnection") as m_conn:
         m_conn.return_value = mock.AsyncMock()
         m_conn.return_value.__aenter__.return_value = m_conn.return_value
         await truncate_and_store_workouts([{"id": 1, "name": "Workout 1"}])
@@ -68,15 +68,3 @@ async def test_truncate_and_store_workouts():
             [{"id": 1, "name": "Workout 1"}]
         )
 
-
-@pytest.mark.asyncio
-async def test_pull_workouts_page():
-    """Test pulling a single page of workouts."""
-    with mock.patch("app.app.httpx.AsyncClient", new=mock.AsyncMock()) as m_client:
-        m_client.get.return_value.json.return_value = {
-            "workouts": [{"id": 1, "name": "Workout 1"}],
-            "page_size": 1,
-            "page_count": 1,
-            "page": 1,
-        }
-        await _pull_workouts_page(1, 1)
