+++ test/test_db_conn.py
@@ -2,12 +2,12 @@
 
 import pytest
 
-from app.conn import Mongo_Conn
+from app.db.conn import MongoConnection
 
 
 # test Mongo_Conn as a context manager
 @pytest.mark.asyncio
 async def test_mongo_conn():
     """Test MongoDB connection context manager."""
-    async with Mongo_Conn("test_db") as db:
+    async with MongoConnection("test_db") as db:
         assert db.name == "test_db"
