+++ test/test_intg_db_conn.py
@@ -2,14 +2,14 @@
 
 import pytest
 
-from app.conn import Mongo_Conn
+from app.db.conn import MongoConnection
 
 
 @pytest.mark.integration
 @pytest.mark.asyncio
 async def test_mongo_conn():
     """Test MongoDB connection context manager."""
-    async with Mongo_Conn("test_db") as db:
+    async with MongoConnection("test_db") as db:
         try:
             await db.create_collection("test_collection")
             assert db.name == "test_db"
